Add patches for Google I/O Schedule
THIS FILE IS NOT A CONTRIBUTION. You may use this file for testing and
evaluation purposes only. Facebook reserves all rights not expressly granted.

This patch set can be applied to The iosched app by running
git checkout 423df831d4bd && tail -n +20 iosched-patches.mbox | git am

It is similar to the patches that enable Buck and exopackage support in
AntennaPod.